Initial Release for Analysis and Comment
NOTE;  This game was made a long time ago and I chose it from a list as the best case example of our talent here at 9 Studio's.  Some things in the game can be upsetting, and we are ROUTING OUT the use of the controversial player "The Mahdi".  That will no longer be and will be replaced by the EU, keeping the play simple as intended.  All files are from an archive of OLD, OLD data and will be used but not as shown.