I find this is the only folder this goes in
I know it's a strong meme for some to hear, but I just wanted to say it as I am in development of a game about this man.  I only want everyone to know, in advance of the game, that if you cannot trust the news coming into your world anymore, there is a problem.  And we should accept it and realize it and fight it.  Social media can't tell you what the history books preserve, even the Muslim Ulama are separated from the flock, while the Muslims are left to Imams who are not doing wrong, but are on a talent show escapade with Social Media apparati.  And we prove this meme to be true daily as the Muslim enemy mocks them in their agitation.  I study media, I see all the attacks, it's too patterned and convenient to just be some mob mentality thing, which it's turning into.  But IMPLIED judgment doesn't mean ACTUAL judgment, and if only we could understand this, when the truth comes out, we won't just follow Tommy Robinson and blindly follow him instead of the truth.  People who are enemies on a dime, and I'm sorry to say that about people who accidentally do and would wish they wouldn't, are not your friends, but everyone is worth love.  Don't give them a reason to turn against you at any moment, help them understand!