Founders Batch
NOTE:  This is satire, we address that there is no real health concern however truthful.  Our founder is adamant that while certain things keep taking place, it makes the imperative nature of this repository more important.  Please do not worry for him, as he asks, he is more than capable of escaping an emergency.